Separated PlayerController and the items
Made PlayerController and Items partial classes. No gameplay change, just for organisation.
