+++ surrounded/Assets/Scripts/Player/PlayerController.cs
@@ -5,7 +5,7 @@ using TMPro;
 using Unity.VisualScripting;
 using Object = UnityEngine.Object;
 
-public class PlayerController : MonoBehaviour
+public partial class PlayerController : MonoBehaviour
 {
     public Rigidbody2D rb;
     public Weapon weapon;
@@ -126,92 +126,5 @@ public class PlayerController : MonoBehaviour
         sprite.color = Color.red;
         yield return new WaitForSeconds(0.1f);
         sprite.color = Color.white;
-    }
-    
-    public void MachineGuns(){ //if machine guns upgrade collected
-        if(MachineGunCount == 0){
-            weapon.fireRate *= 1.1f;
-        } else{
-            weapon.fireRate = (1.1f+(0.05f*MachineGunCount)) * weapon.baseRate;
-        }
-        MachineGunCount++;
-        }
-
-    public void RocketBooster(){
-        if(RocketBoosterCount == 0){
-            moveSpeed *= 1.1f;
-        } else{
-            moveSpeed = (1.1f+(0.05f*RocketBoosterCount)) * baseSpeed;
-        }
-        RocketBoosterCount++;
-    }
-
-    public void PilotingEnhancement(){
-        XP += levelReq;
-        XP += levelReq;
-    }
-    
-    public void DivergeActivated(){
-        divergeActivated = true;
-        divergeCount++;
-    }
-
-    public void Stats(){
-        stats.text = "ATK: " + (int)damage + " DEF: " + (int)defense + " SPD: " + (int)moveSpeed;
-        sstats.text = "ATK: " + (int)damage + " DEF: " + (int)defense + " SPD: " + (int)moveSpeed;
-    }
-
-    public void Shield(){   
-        maxShield = (0.1f+(0.05f*shieldCount))*maxHealth;
-        shieldCount++;
-        shield = maxShield;
-        moveSpeed = (1-(0.05f*shieldCount))*moveSpeed;
-    }
-
-    public void ForceField()
-    {
-        forceFieldActivated = true;
-        forcefieldCount++;
-    }
-
-    public IEnumerator forceFieldTimer(int totaltime)
-    {
-        while (_timer < totaltime)
-        {
-            yield return new WaitForSecondsRealtime(1);
-            _timer++;
-            Debug.Log("forcefield timer at " + _timer);
-        }
-        
-        // trigger callback
-        onTimeOut?.Invoke();
-    }
-
-    public void ForceFieldTimerStart(int totalTime, Action timeOut)
-    {
-        onTimeOut = timeOut; // save callback Action
-        // reset timer
-        _timer = 0;
-        TimerCoroutine = forceFieldTimer(totalTime);
-        StartCoroutine(TimerCoroutine);
-    }
-    public void Roulette(){
-        rouletteCount++;
-        if(rouletteCount > 3){
-            rouletteCount = 3;
-        }
-        GameObject roulette = Instantiate(rouletteBall);
-        Roulette r = roulette.GetComponent<Roulette>();
-        switch(rouletteCount){
-            case 1: 
-                r.isBall1 = true;
-                break;
-            case 2:
-                r.isBall2 = true;
-                break;
-            case 3:
-                r.isBall3 = true;
-                break;
-        }
     }
 }
